Improve logging consistency and verbosity

+++ src/Kongverge/Workflow/KongvergeWorkflow.cs
@@ -7,6 +7,7 @@ using Kongverge.DTOs;
 using Kongverge.Helpers;
 using Kongverge.Services;
 using Microsoft.Extensions.Options;
+using Serilog;
 
 namespace Kongverge.Workflow
 {
@@ -45,8 +46,13 @@ namespace Kongverge.Workflow
             }
 
             var existingConfiguration = await _configBuilder.FromKong(KongReader);
+
+            await ConvergeChildrenPlugins(null, existingConfiguration.GlobalConfig, targetConfiguration.GlobalConfig);
             
             await ConvergeObjects(
+                null,
+                "service",
+                "services",
                 existingConfiguration.Services,
                 targetConfiguration.Services,
                 x => _kongWriter.DeleteService(x.Id),
@@ -54,28 +60,41 @@ namespace Kongverge.Workflow
                 x => _kongWriter.UpdateService(x),
                 ConvergeServiceChildren);
 
-            await ConvergeChildrenPlugins(existingConfiguration.GlobalConfig, targetConfiguration.GlobalConfig);
-
             return ExitWithCode.Return(ExitCode.Success);
         }
 
         private static async Task ConvergeObjects<T>(
+            string parent,
+            string singularObjectName,
+            string pluralObjectName,
             IReadOnlyCollection<T> existingObjects,
             IReadOnlyCollection<T> targetObjects,
             Func<T, Task> deleteObject,
             Func<T, Task> createObject,
             Func<T, Task> updateObject = null,
-            Func<T, T, Task> recurse = null) where T : KongObject
+            Func<T, T, Task> recurse = null) where T : KongObject, IKongEquatable
         {
             existingObjects = existingObjects ?? Array.Empty<T>();
             updateObject = updateObject ?? (x => Task.CompletedTask);
             recurse = recurse ?? ((e, t) => Task.CompletedTask);
 
+            if (existingObjects.Count == 0 && targetObjects.Count == 0)
+            {
+                Log.Information($"Target {parent ?? "configuration"} and existing both have zero {pluralObjectName}");
+                return;
+            }
+
+            var targetPhrase = $"{targetObjects.Count} target {GetName(targetObjects.Count, singularObjectName, pluralObjectName)}";
+            var existingPhrase = $"{existingObjects.Count} existing {GetName(existingObjects.Count, singularObjectName, pluralObjectName)}";
+            var parentPhrase = parent == null ? string.Empty : $" attached to {parent}";
+            Log.Information($"Converging {targetPhrase}{parentPhrase} with {existingPhrase}");
+
             var targetMatchValues = targetObjects.Select(x => x.GetMatchValue()).ToArray();
-            var toRemove = existingObjects.Where(x => !targetMatchValues.Contains(x.GetMatchValue()));
+            var toRemove = existingObjects.Where(x => !targetMatchValues.Contains(x.GetMatchValue())).ToArray();
 
             foreach (var existing in toRemove)
             {
+                Log.Information($"Deleting {singularObjectName} {existing}{parentPhrase} which exists in Kong but not in target configuration");
                 await deleteObject(existing);
             }
 
@@ -84,25 +103,35 @@ namespace Kongverge.Workflow
                 var existing = target.MatchWithExisting(existingObjects);
                 if (existing == null)
                 {
+                    Log.Information($"Creating {singularObjectName} {target}{parentPhrase} which exists in target configuration but not in Kong");
                     await createObject(target);
                 }
-                else if (!target.Equals(existing))
+                else if (target.Equals(existing))
                 {
+                    Log.Information($"Identical {singularObjectName} {existing}{parentPhrase} found in Kong matching target configuration");
+                }
+                else
+                {
+                    var patch = target.DifferencesFrom(existing);
+                    Log.Information($"Updating {singularObjectName} {existing}{parentPhrase} which exists in both Kong and target configuration, having the following differences:{Environment.NewLine}{patch}");
                     await updateObject(target);
                 }
                 await recurse(existing, target);
             }
         }
 
-        private Task ConvergeChildrenPlugins(IKongPluginHost existing, IKongPluginHost target)
+        private Task ConvergeChildrenPlugins(string parent, IKongPluginHost existing, IKongPluginHost target)
         {
-            Task UpsertPlugin(KongPlugin plugin, IKongPluginHost parent)
+            Task UpsertPlugin(KongPlugin plugin, IKongPluginHost host)
             {
-                parent.AssignParentId(plugin);
+                host.AssignParentId(plugin);
                 return _kongWriter.UpsertPlugin(plugin);
             }
 
             return ConvergeObjects(
+                parent,
+                "plugin",
+                "plugins",
                 existing?.Plugins,
                 target.Plugins,
                 x => _kongWriter.DeletePlugin(x.Id),
@@ -112,14 +141,25 @@ namespace Kongverge.Workflow
 
         private async Task ConvergeServiceChildren(KongService existing, KongService target)
         {
-            await ConvergeChildrenPlugins(existing, target);
+            var parent = $"service {target}";
+            await ConvergeChildrenPlugins(parent, existing, target);
             await ConvergeObjects(
+                parent,
+                "route",
+                "routes",
                 existing?.Routes,
                 target.Routes,
                 x => _kongWriter.DeleteRoute(x.Id),
                 x => _kongWriter.AddRoute(target.Id, x),
                 null,
-                ConvergeChildrenPlugins);
+                (e, t) => ConvergeChildrenPlugins($"route {t} attached to service {target}", e, t));
+        }
+
+        private static string GetName(int count, string singular, string plural)
+        {
+            return count == 1
+                ? singular
+                : plural;
         }
     }
 }
